requirements.txt
mitmproxy
pyyaml